requirements.txt
# aeroRAG - NASA STI Repository RAG System
# Environment Setup Requirements

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Text Processing & Chunking
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Embeddings (November Phase)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # For vector storage and similarity search
chromadb>=0.4.0  # Alternative vector database

# LLM Integration (Ollama)
ollama>=0.1.0
langchain-community>=0.0.20  # For Ollama integration

# Web Interface
streamlit>=1.28.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0  # Progress bars
python-dotenv>=1.0.0  # Environment variables

# Optional: For advanced processing
scikit-learn>=1.3.0  # For clustering/analysis
matplotlib>=3.7.0  # For visualization
seaborn>=0.12.0  # For visualization

